Newest1-main/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncssh>=2.20.0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "motor>=3.7.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "py-cord>=2.4.0",
    "python-dateutil>=2.9.0.post0",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.40",
    "asyncpg>=0.30.0",
    "discord-py>=2.5.2",
    "pymongo>=4.12.0",
]
